ApacheKafkaInPython/requirements.txt
pip~=23.2.1
kafka-python